#19 Remove authenticator from handler tests
