+++ presentation/handler/user_handler_test.go
@@ -3,50 +3,15 @@ package handler
 import (
 	"bytes"
 	"encoding/json"
-	"github.com/gofiber/fiber/v2"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"io"
 	"net/http"
-	"os"
-	"split-the-bill-server/authentication"
 	"split-the-bill-server/domain/service/mocks"
-	"split-the-bill-server/domain/service/service_inf"
 	"split-the-bill-server/presentation/dto"
 	. "split-the-bill-server/storage/database/test_util"
 	"testing"
 )
-
-var (
-	userService service_inf.IUserService
-	userHandler UserHandler
-	app         *fiber.App
-)
-
-func TestMain(m *testing.M) {
-	// setup user handler
-	userService = mocks.NewUserServiceMock()
-	// password validator
-	passwordValidator, err := authentication.NewPasswordValidator()
-	if err != nil {
-		panic("Error while setting up the password validator: " + err.Error())
-	}
-	userHandler = *NewUserHandler(&userService, passwordValidator)
-
-	// TODO: setup authentication handler
-
-	// setup fiber
-	app = fiber.New()
-	// TODO: use SetupRoutes
-	app.Get("/user/:id", userHandler.GetByID)
-	app.Post("/api/user/register", userHandler.Register)
-
-	// Run tests
-	exitCode := m.Run()
-
-	// Exit with the same code as the test run
-	os.Exit(exitCode)
-}
 
 func TestGetByIDSuccess(t *testing.T) {
 
@@ -57,18 +22,18 @@ func TestGetByIDSuccess(t *testing.T) {
 
 	// setup request
 	req, _ := http.NewRequest(http.MethodGet, "/user/"+User.ID.String(), nil)
-	resp, err := app.Test(req, -1)
+	resp, err := app.Test(req)
 	if resp != nil {
 		defer resp.Body.Close()
 	}
 	if err != nil {
-		panic(err)
+		t.Fatal("Error in test setup during performing a request - ", err)
 	}
 
 	// read response body
 	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		panic(err)
+		t.Fatal("Error in test setup during reading response body - ", err)
 	}
 	var response dto.GeneralResponseDTO
 	err = json.Unmarshal(body, &response)
@@ -103,18 +68,17 @@ func TestRegisterSuccess(t *testing.T) {
 	jsonBody, err := json.Marshal(reqBody)
 	req, _ := http.NewRequest(http.MethodPost, "/api/user/register", bytes.NewBuffer(jsonBody))
 	req.Header.Set("Content-Type", "application/json")
-	resp, err := app.Test(req, -1)
+	resp, err := app.Test(req)
 	if resp != nil {
 		defer resp.Body.Close()
 	}
 	if err != nil {
-		panic(err)
+		t.Fatal("Error in test setup during performing a request - ", err)
 	}
-
 	// read response body
 	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		panic(err)
+		t.Fatal("Error in test setup during reading response body - ", err)
 	}
 	var response dto.GeneralResponseDTO
 	err = json.Unmarshal(body, &response)
